Master Duel Banlist
yo its me master duel man and im going to be in control of the master duel banlist i thank the discord server for providing me with the current list for me to post now and i will update it on regular basis. i want to do this so people can test decks with confirmed new cards coming to master duel. i have an rca cambio i can use to update the list with so dvs please incorperate it into project ignis thank you